refactor(utils): move isEqual in utils folder
