+++ src/Dropdown/Dropdown.jsx
@@ -20,6 +20,7 @@ import Option from './Option'
 import Heading from './Heading'
 import sizes from '../assets/sizes'
 import { SubLabel } from '../Input/SubLabel'
+import { isEqual } from '../utils/isEqual'
 
 const Dropdown = ({
   disabled,
@@ -320,31 +321,6 @@ const Dropdown = ({
       />
     </StyledFlex>
   )
-}
-
-const isEqual = (object1, object2) => {
-  if (typeof object1 !== typeof object2) {
-    return false
-  }
-
-  if (typeof object1 !== 'object') {
-    return object1 === object2
-  }
-
-  const keys1 = Object.keys(object1 ?? {})
-  const keys2 = Object.keys(object2 ?? {})
-
-  if (keys1.length !== keys2.length) {
-    return false
-  }
-
-  for (let key of keys1) {
-    if (!isEqual(object1[key], object2[key])) {
-      return false
-    }
-  }
-
-  return true
 }
 
 const StyledFlex = styled(Flex)(
